Auto-detect whether the tests are running in-app.
The tests were written to run as InApp, but InApp excludes everything
that has "vendor" in its import path. This only becomes a problem when
raven-go is vendored, and you run `go test ./...` for the package. On
the one hand, the package can be tested separately. On the other, it's
nice to know that all your dependencies pass their tests, and it's
pretty trivial to support making this pass when vendored.

Basically, we just need to update the test to detect when it's running
in a situation that's not considered "InApp", and then make sure that
the Stacktrace reports that correctly. Fortunately, we know how to
decide if something is InApp: does it have "vendor" in its path?

So we just check if the current file (retrieved using runtime.Caller)
has "vendor" in its path, and use that to decide whether the test passes
or fails, instead of the hardcoded `true`.