stacktrace.go: fileContext: Cache errors to avoid repeated syscalls
If the source files are missing or owned by the wrong user, every
call to NewStacktrace causes an attempt to open and read the file.
This causes wasted work on servers that are deployed without their
source code.

Change fileContext to cache errors as an empty file (nil slice). This
changes this function to only try to read each unique file once.
Fixes issue #128.