Client.Capture: Return closed chan for nil clients.
The API docs state that Capture "is a no-op when client is nil." This
suggests that there should be no errors when calling it, so receiving
an error from the error chan result should immediately return nil.
Previously, it would return a nil chan. Receiving from a nil chan
blocks forever, which caused callers to need to check for nil clients
in some cases.

Change Capture to return a closed chan for nil clients, which will
immediately return nil.