Add methods that return error when capturing
The client's capture method returns an error channel that according to
the docs is intended to be used for checking if a packet was sent
successfully whenever that is important.

The `...AndWait` methods use this channel, but only to wait. They do
not capture the possible error coming from that channel.

The changes in this commit suggest to use the error and return it, so
that a user may check whether a packet was sent successfully using new top
level methods, instead of having to write their own.

The newly introduced methods follow the naming scheme
`Capture$SOMETHINGAndConfirm`, where the returned error is considered
confirmation.

Work on #84