Capture and CaptureError not to panic when packet or err is nil
Capture methods can handle situations when client is nil, but when
packet or err args are nil it will cause calling thread to panic.
This is highly undesirable in PROD systems that accidentally try to log
a nil packet or error as it might cause a system crash on something
not produciton critical as error logging.